Move out style helpers and add test
