+++ lighthouse-core/test/audits/dobetterweb/uses-new-flexbox-test.js
@@ -1,0 +1,49 @@
+/**
+ * Copyright 2016 Google Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+'use strict';
+
+const UsesNewFlexBoxAudit = require('../../../audits/dobetterweb/uses-new-flexbox.js');
+const assert = require('assert');
+const stylesData = require('../../fixtures/styles-gatherer.json');
+
+/* eslint-env mocha */
+
+describe('Page does not use old CSS flexbox', () => {
+  it('fails when no input present', () => {
+    const auditResult = UsesNewFlexBoxAudit.audit({});
+    assert.equal(auditResult.rawValue, -1);
+    assert.ok(auditResult.debugString);
+  });
+
+  it('fails when display: box is used', () => {
+    const auditResult = UsesNewFlexBoxAudit.audit({
+      Styles: stylesData
+    });
+    assert.equal(auditResult.rawValue, false);
+    assert.equal(auditResult.extendedInfo.value.length, 2);
+    assert.equal(auditResult.extendedInfo.value.length, 2);
+    assert.equal(auditResult.extendedInfo.value[0].url, stylesData[0].header.sourceURL);
+    assert.ok(auditResult.extendedInfo.value[0].misc.match(/display\: box/));
+  });
+
+  it('passes when display: box is not used', () => {
+    const auditResult = UsesNewFlexBoxAudit.audit({
+      Styles: stylesData.slice(1)
+    });
+    assert.equal(auditResult.rawValue, true);
+    assert.equal(auditResult.extendedInfo.value.length, 0);
+  });
+});
